backend/pyproject.toml
[tool.poetry]
authors = ["Mert Kayhan <mertkayhan.1993@gmail.com>"]
description = ""
name = "jarvis"
readme = "README.md"
version = "0.0.1"

[tool.poetry.dependencies]
PyMuPDF = "==1.25.3"
aiohttp = "==3.11.13"
auth0-python = "==4.8.1"
eventlet = "==0.35.2"
fastapi = {version = "==0.115.12", extras = ["standard"]}
gcsfs = "==2024.10.0"
graphrag = "==2.1.0"
langchain = "==0.3.17"
langchain-anthropic = "==0.3.6"
langchain-google-vertexai = "==2.0.13"
langchain-openai = "==0.3.3"
langgraph = "==0.2.69"
langgraph-checkpoint-postgres = "==2.0.10"
llama-parse = "==0.6.4.post1"
markdown = "==3.7"
markdownify = "==0.14.1"
numpy = "==1.26.4"
pillow = "==10.2.0"
python = ">=3.10,<3.13"
python-dotenv = "^1.0.0"
python-socketio = "==5.11.1"
ragas = "==0.2.8"
setuptools = "==71.1.0"
tabulate = "==0.9.0"
tavily-python = "==0.5.0"
unstructured = "==0.17.2"

[tool.poetry.dev-dependencies]
black = "==24.2.0"
mypy = "==1.15.0"

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
  "google.cloud",
  "langchain_core.*",
  "langchain_openai",
  "psycopg_pool",
  "tavily",
  "psycopg.*",
]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
